+++ scout/server/blueprints/cases/views.py
@@ -313,6 +313,22 @@ def pdf_case_report(institute_id, case_name):
     )
 
 
+@cases_bp.route("/<institute_id>/<case_name>/chanjo2_coverage_report", methods=["GET"])
+def chanjo2_coverage_report(institute_id: str, case_name: str):
+    """Return the HTML coverage report created by chanjo2."""
+
+    institute_obj, case_obj = institute_and_case(store, institute_id, case_name)
+    report_html_content: str = controllers.chanjo2_coverage_report_contents(
+        institute_obj=institute_obj,
+        case_obj=case_obj,
+        panel_name=request.args.get("panel_name"),
+        panel_id=request.args.get("panel_id"),
+    )
+    if report_html_content is None:
+        return redirect(request.referrer)
+    return report_html_content
+
+
 @cases_bp.route("/<institute_id>/<case_name>/mt_report", methods=["GET"])
 def mt_report(institute_id, case_name):
     institute_obj, case_obj = institute_and_case(store, institute_id, case_name)