+++ CHANGELOG.md
@@ -5,12 +5,16 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 About changelog [here](https://keepachangelog.com/en/1.0.0/)
 
 ## [unreleased]
+### Added
+- Support for .d4 files coverage using chanjo2 (Case page sidebar link) with test
+- Link for chanjo2 coverage report on gene panel page
 ### Changed
 - All links in disease table on diagnosis page now open in a new tab
 - Dark mode settings applied to multiselects on institute settings
 - Comments on case and variant pages can be viewed by expanding an accordion
 ### Fixed
 - On variants page, search for variants in genes present only in build 38 returning no results
+- Pin/unpin with API was not able to make event links
 
 ## [4.79.1]
 ### Fixed